refactor: decompose ApiServiceProvider into registrar collaborators
Extract the macro, middleware, logging, lifecycle, and container
binding registration blocks into dedicated final registrars under
Providers\Registrars, leaving the provider as a thin coordinator.
Registration order, config keys, and resolution timing are preserved
exactly; no public API changes. Clears the final S1448 finding.
